// src/app/layouts/login-page/model/pinjaman.ts
export class PinjamanApproval {
  id: any;
  idNasabah: any;
  nominalTransaksi: any;
  tanggal: any;
  waktu: any;
  bulanBayar: any;
  tujuanPinjam: any;
  nip: any;
  namaNasabah: any;
}


export class PinjamanParameter {
  no: any;
  id: any;
  idNasabah: any;
  nominalPinjaman: any;
  nominalTransaksi: any;
  bulanBayar: any;
  noPinjaman: any;
  statusTransaksi: any;
  buktiPembayaran: any;
  tanggalTransaksi: any;
  tanggalApprove: any;
  adminAction: any;
  adminApprove: any;
  alasanTolak: any;
  deskripsi: any;
  idApproval: any;
  tujuanPinjam: any;
  nip: any;
  namaNasabah: any;
}

export class Pinjaman {

  id : any;
  idNasabah : any;
  nominalPinjaman : any;
  noPinjaman : any;
  statusTransaksi : any;
  buktiPembayaran : any;
  tanggalTransaksi : any;
  deskripsi : any;
}